ref(server): Move project source into the source module
